+++ participants/src/Piipan.Participants/Piipan.Participants.Core/Models/UploadDbo.cs
@@ -5,7 +5,7 @@ namespace Piipan.Participants.Core.Models
 {
     public class UploadDbo : IUpload
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public DateTime CreatedAt { get; set; }
         public string Publisher { get; set; }
     }
